pyproject.toml
[tool.poetry]
name = "creampie"
version = "0.1.0"
description = ""
authors = ["Robert Ferguson <rmferguson@protonmail.com>"]
readme = "cream.md"
package-mode = true
packages = [
    { include = "cream_api" }
]

[tool.poetry.dependencies]
python = "^3.12"
cryptograpy = "^0.0.0"
llama-cpp-python = "^0.3.9"
numpy = "^2.3.0"
pandas = "^2.3.0"
requests = "^2.32.4"
scipy = "^1.15.3"
seleniumbase = "^4.39.4"
stargazer-utils = "^1.0.0"
fastapi = {extras = ["standard"], version = "^0.115.12"}
sqlalchemy = "^2.0.41"
pydantic-settings = "^2.9.1"
psycopg = "^3.2.9"
alembic = "^1.13.1"
aiohttp = "^3.12.13"
beautifulsoup4 = "^4.13.4"
jwt = "^1.3.1"


[tool.poetry.group.dev.dependencies]
autoflake = "^2.3.1"
hypothesis = "^6.135.9"
mypy = "^1.16.0"
pandas-stubs = "^2.2.3.250527"
pylint = "^3.3.7"
pytest = "^8.4.0"
scipy-stubs = "^1.15.3.0"
types-requests = "^2.32.4.20250611"
swent = "^0.1.0"
tombi = "^0.4.9"
ruff = "^0.11.13"
coverage = "^7.9.1"
click = "^8.2.1"
pre-commit = "^4.2.0"
pytest-asyncio = "^1.0.0"
types-beautifulsoup4 = "^4.12.0.20250516"
aiosqlite = "^0.21.0"
aioresponses = "^0.7.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
exclude = [
    'tests/.*',
    '.*_test\.py$',
    'test_.*\.py$'
]

[[tool.mypy.overrides]]
module = [
    "aiohttp.*",
    "pytest_asyncio.*",
    "pytest.*",
    "hypothesis.*",
    "aioresponses.*"
]
ignore_missing_imports = true
